petstore-api-tests/requirements.txt
pytest==8.3.4
requests==2.32.3
pytest-html==4.1.1
pytest-xdist==3.6.1
python-dotenv==1.0.1
faker==33.1.0
